Clean up Alexa smart home code
