+++ homeassistant/components/alexa/__init__.py
@@ -5,12 +5,13 @@ import voluptuous as vol
 
 from homeassistant.helpers import config_validation as cv
 from homeassistant.helpers import entityfilter
+from homeassistant.const import CONF_NAME
 
-from . import flash_briefings, intent, smart_home
+from . import flash_briefings, intent, smart_home_http
 from .const import (
     CONF_AUDIO, CONF_CLIENT_ID, CONF_CLIENT_SECRET, CONF_DISPLAY_URL,
     CONF_ENDPOINT, CONF_TEXT, CONF_TITLE, CONF_UID, DOMAIN, CONF_FILTER,
-    CONF_ENTITY_CONFIG)
+    CONF_ENTITY_CONFIG, CONF_DESCRIPTION, CONF_DISPLAY_CATEGORIES)
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -18,9 +19,9 @@ CONF_FLASH_BRIEFINGS = 'flash_briefings'
 CONF_SMART_HOME = 'smart_home'
 
 ALEXA_ENTITY_SCHEMA = vol.Schema({
-    vol.Optional(smart_home.CONF_DESCRIPTION): cv.string,
-    vol.Optional(smart_home.CONF_DISPLAY_CATEGORIES): cv.string,
-    vol.Optional(smart_home.CONF_NAME): cv.string,
+    vol.Optional(CONF_DESCRIPTION): cv.string,
+    vol.Optional(CONF_DISPLAY_CATEGORIES): cv.string,
+    vol.Optional(CONF_NAME): cv.string,
 })
 
 SMART_HOME_SCHEMA = vol.Schema({
@@ -65,6 +66,6 @@ async def async_setup(hass, config):
         pass
     else:
         smart_home_config = smart_home_config or SMART_HOME_SCHEMA({})
-        await smart_home.async_setup(hass, smart_home_config)
+        await smart_home_http.async_setup(hass, smart_home_config)
 
     return True
