+++ homeassistant/components/cloud/__init__.py
@@ -4,11 +4,11 @@ import logging
 import voluptuous as vol
 
 from homeassistant.auth.const import GROUP_ID_ADMIN
-from homeassistant.components.alexa import smart_home as alexa_sh
+from homeassistant.components.alexa import const as alexa_const
 from homeassistant.components.google_assistant import const as ga_c
 from homeassistant.const import (
     CONF_MODE, CONF_NAME, CONF_REGION, EVENT_HOMEASSISTANT_START,
-    EVENT_HOMEASSISTANT_STOP)
+    EVENT_HOMEASSISTANT_STOP, CONF_NAME)
 from homeassistant.core import callback
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import config_validation as cv, entityfilter
@@ -33,9 +33,9 @@ SERVICE_REMOTE_DISCONNECT = 'remote_disconnect'
 
 
 ALEXA_ENTITY_SCHEMA = vol.Schema({
-    vol.Optional(alexa_sh.CONF_DESCRIPTION): cv.string,
-    vol.Optional(alexa_sh.CONF_DISPLAY_CATEGORIES): cv.string,
-    vol.Optional(alexa_sh.CONF_NAME): cv.string,
+    vol.Optional(alexa_const.CONF_DESCRIPTION): cv.string,
+    vol.Optional(alexa_const.CONF_DISPLAY_CATEGORIES): cv.string,
+    vol.Optional(CONF_NAME): cv.string,
 })
 
 GOOGLE_ENTITY_SCHEMA = vol.Schema({
